Small refactoring
Moved from usize to u64 in most places.
Added working support for sorting by rate

+++ src/file.rs
@@ -5,14 +5,14 @@ use std::sync::{Arc, Mutex};
 
 /// Saves result hashmap into given file.
 pub fn save_result(
-    map: Arc<Mutex<HashMap<String, usize>>>,
+    map: Arc<Mutex<HashMap<String, u64>>>,
     filename: String,
 ) -> Result<(), std::io::Error> {
     let map = map.lock().expect("Unable to lock mutex.");
     let file = File::create(filename)?;
     let mut file = BufWriter::new(file);
 
-    let mut sorted: Vec<(&String, &usize)> = map.iter().collect();
+    let mut sorted: Vec<(&String, &u64)> = map.iter().collect();
     sorted.sort_by(|a, b| b.1.cmp(a.1));
 
     for (k, v) in sorted {
